+++ src/code42cli/cmds/profile.py
@@ -5,6 +5,7 @@ from click import echo
 from click import secho
 
 import code42cli.profile as cliprofile
+from code42cli.click_ext.options import incompatible_with
 from code42cli.click_ext.types import PromptChoice
 from code42cli.click_ext.types import TOTP
 from code42cli.errors import Code42CLIError
@@ -58,12 +59,14 @@ def username_option(required=False):
         "-u",
         "--username",
         required=required,
+        cls=incompatible_with(["api_client_id", "secret"]),
         help="The username of the Code42 API user.",
     )
 
 
 password_option = click.option(
     "--password",
+    cls=incompatible_with(["api_client_id", "secret"]),
     help="The password for the Code42 API user. If this option is omitted, interactive prompts "
     "will be used to obtain the password.",
 )
@@ -84,18 +87,44 @@ use_v2_file_events_option = click.option(
 )
 
 
+def api_client_id_option(required=False):
+    return click.option(
+        "--api-client-id",
+        required=required,
+        cls=incompatible_with(["username", "password", "totp"]),
+        help="The API client key for API client authentication.  Used with the `--secret` option.",
+    )
+
+
+def secret_option(required=False):
+    return click.option(
+        "--secret",
+        required=required,
+        cls=incompatible_with(["username", "password", "totp"]),
+        help="The API secret for API client authentication.  Used with the `--api-client` option.",
+    )
+
+
 @profile.command()
 @profile_name_arg()
 def show(profile_name):
     """Print the details of a profile."""
     c42profile = cliprofile.get_profile(profile_name)
     echo(f"\n{c42profile.name}:")
-    echo(f"\t* username = {c42profile.username}")
+    if c42profile.api_client_auth == "True":
+        echo(f"\t* api-client-id = {c42profile.username}")
+    else:
+        echo(f"\t* username = {c42profile.username}")
     echo(f"\t* authority url = {c42profile.authority_url}")
     echo(f"\t* ignore-ssl-errors = {c42profile.ignore_ssl_errors}")
     echo(f"\t* use-v2-file-events = {c42profile.use_v2_file_events}")
-    if cliprofile.get_stored_password(c42profile.name) is not None:
-        echo("\t* A password is set.")
+    echo(f"\t* api-client-auth-profile = {c42profile.api_client_auth}")
+    if c42profile.api_client_auth == "True":
+        if cliprofile.get_stored_password(c42profile.name) is not None:
+            echo("\t* The API client secret is set.")
+    else:
+        if cliprofile.get_stored_password(c42profile.name) is not None:
+            echo("\t* A password is set.")
     echo("")
     echo("")
 
@@ -120,28 +149,75 @@ def create(
     debug,
     totp,
 ):
-    """Create profile settings. The first profile created will be the default."""
+    """
+    Create a profile with username/password authentication.
+    The first profile created will be the default.
+    """
     cliprofile.create_profile(
-        name, server, username, disable_ssl_errors, use_v2_file_events
+        name,
+        server,
+        username,
+        disable_ssl_errors,
+        use_v2_file_events,
+        api_client_auth=False,
     )
-    password = password or _prompt_for_password(name)
+    password = password or _prompt_for_password()
     if password:
-        _set_pw(name, password, debug, totp=totp)
+        _set_pw(name, password, debug, totp=totp, api_client=False)
+    echo(f"Successfully created profile '{name}'.")
+
+
+@profile.command()
+@name_option(required=True)
+@server_option(required=True)
+@api_client_id_option(required=True)
+@secret_option(required=True)
+@yes_option(hidden=True)
+@disable_ssl_option
+@use_v2_file_events_option
+@debug_option
+def create_api_client(
+    name,
+    server,
+    api_client_id,
+    secret,
+    disable_ssl_errors,
+    use_v2_file_events,
+    debug,
+):
+    """
+    Create a profile with Code42 API client authentication.
+    The first profile created will be the default.
+    """
+    cliprofile.create_profile(
+        name,
+        server,
+        api_client_id,
+        disable_ssl_errors,
+        use_v2_file_events,
+        api_client_auth=True,
+    )
+    _set_pw(name, secret, debug, totp=False, api_client=True)
     echo(f"Successfully created profile '{name}'.")
 
 
 @profile.command()
 @name_option()
 @server_option()
+@api_client_id_option()
+@secret_option()
 @username_option()
 @password_option
 @totp_option
 @disable_ssl_option
 @use_v2_file_events_option
+@yes_option(hidden=True)
 @debug_option
 def update(
     name,
     server,
+    api_client_id,
+    secret,
     username,
     password,
     disable_ssl_errors,
@@ -152,24 +228,102 @@ def update(
     """Update an existing profile."""
     c42profile = cliprofile.get_profile(name)
 
-    if (
-        not server
-        and not username
-        and not password
-        and disable_ssl_errors is None
-        and use_v2_file_events is None
-    ):
-        raise click.UsageError(
-            "Must provide at least one of `--username`, `--server`, `--password`, `--use-v2-file-events` or "
-            "`--disable-ssl-errors` when updating a profile."
-        )
-    cliprofile.update_profile(
-        c42profile.name, server, username, disable_ssl_errors, use_v2_file_events
-    )
-    if not password and not c42profile.has_stored_password:
-        password = _prompt_for_password(c42profile.name)
-    if password:
-        _set_pw(name, password, debug, totp=totp)
+    if c42profile.api_client_auth == "True":
+        if not any(
+            [
+                server,
+                api_client_id,
+                secret,
+                disable_ssl_errors is not None,
+                use_v2_file_events is not None,
+            ]
+        ):
+            raise click.UsageError(
+                "Must provide at least one of `--server`, `--api-client-id`, `--secret`, `--use-v2-file-events` or "
+                "`--disable-ssl-errors` when updating an API client profile.  "
+                "Provide both `--username` and `--password` options to switch this profile to username/password authentication."
+            )
+        if (username and not password) or (password and not username):
+            raise click.UsageError(
+                "This profile currently uses API client authentication.  "
+                "Please provide both the `--username` and `--password` options to update this profile to use username/password authentication."
+            )
+        elif username and password:
+            if does_user_agree(
+                "You passed the `--username` and `--password options for a profile currently using Code42 API client authentication.  "
+                "Are you sure you would like to update this profile to use username/password authentication? This will overwrite existing credentials. (y/n): "
+            ):
+                cliprofile.update_profile(
+                    c42profile.name,
+                    server,
+                    username,
+                    disable_ssl_errors,
+                    use_v2_file_events,
+                    api_client_auth=False,
+                )
+                _set_pw(c42profile.name, password, debug, api_client=False)
+            else:
+                echo(f"Profile '{c42profile.name}` was not updated.")
+                return
+        else:
+            cliprofile.update_profile(
+                c42profile.name,
+                server,
+                api_client_id,
+                disable_ssl_errors,
+                use_v2_file_events,
+            )
+            if secret:
+                _set_pw(c42profile.name, secret, debug, api_client=True)
+
+    else:
+        if (
+            not server
+            and not username
+            and not password
+            and disable_ssl_errors is None
+            and use_v2_file_events is None
+        ):
+            raise click.UsageError(
+                "Must provide at least one of `--server`, `--username`, `--password`, `--use-v2-file-events` or "
+                "`--disable-ssl-errors` when updating a username/password authenticated profile.  "
+                "Provide both `--api-client-id` and `--secret` options to switch this profile to Code42 API client authentication."
+            )
+        if (api_client_id and not secret) or (api_client_id and not secret):
+            raise click.UsageError(
+                "This profile currently uses username/password authentication.  "
+                "Please provide both the `--api-client-id` and `--secret` options to update this profile to use Code42 API client authentication."
+            )
+        elif api_client_id and secret:
+            if does_user_agree(
+                "You passed the `--api-client-id` and `--secret options for a profile currently using username/password authentication.  "
+                "Are you sure you would like to update this profile to use Code42 API client authentication? This will overwrite existing credentials. (y/n): "
+            ):
+                cliprofile.update_profile(
+                    c42profile.name,
+                    server,
+                    api_client_id,
+                    disable_ssl_errors,
+                    use_v2_file_events,
+                    api_client_auth=True,
+                )
+                _set_pw(c42profile.name, secret, debug, api_client=True)
+            else:
+                echo(f"Profile '{name}` was not updated.")
+                return
+        else:
+            cliprofile.update_profile(
+                c42profile.name,
+                server,
+                username,
+                disable_ssl_errors,
+                use_v2_file_events,
+            )
+            if not password and not c42profile.has_stored_password:
+                password = _prompt_for_password()
+
+            if password:
+                _set_pw(c42profile.name, password, debug, totp=totp)
 
     echo(f"Profile '{c42profile.name}' has been updated.")
 
@@ -251,16 +405,22 @@ def delete_all():
         echo("\nNo profiles exist. Nothing to delete.")
 
 
-def _prompt_for_password(profile_name):
+def _prompt_for_password():
     if does_user_agree("Would you like to set a password? (y/n): "):
         password = getpass()
         return password
 
 
-def _set_pw(profile_name, password, debug, totp=None):
+def _set_pw(profile_name, password, debug, totp=None, api_client=False):
     c42profile = cliprofile.get_profile(profile_name)
     try:
-        create_sdk(c42profile, is_debug_mode=debug, password=password, totp=totp)
+        create_sdk(
+            c42profile,
+            is_debug_mode=debug,
+            password=password,
+            totp=totp,
+            api_client=api_client,
+        )
     except Exception:
         secho("Password not stored!", bold=True)
         raise