+++ src/code42cli/options.py
@@ -61,7 +61,12 @@ class CLIState:
     @property
     def sdk(self):
         if self._sdk is None:
-            self._sdk = create_sdk(self.profile, self.debug, totp=self.totp)
+            self._sdk = create_sdk(
+                self.profile,
+                self.debug,
+                totp=self.totp,
+                api_client=self.profile.api_client_auth == "True",
+            )
         return self._sdk
 
     def set_assume_yes(self, param):